refactor: split MainView.swift into AppState, ContentView, and MainView
Extract AppState class and model structs into AppState.swift, and
ContentView/EmptyStateView into ContentView.swift. No behavioral changes.
